bench: use zipfian directly and non-determininstic seeds
The purpose of the scrambling generators is to move keys around
so that hot keys are spatially distributed.
What happened before is that when new zipfian generator was made,
the set of keys that were considered hot was always new.
This is problematic for multi-step traces and warmups.

We tried to fix this by making initialization deterministic,
but this results in each step in a multi-step trace getting the
same keys.

Since keys are scrambled anyway due to fnv hashing,
scrambling in the integer generators was redundant anyway.
Just drop the scrambling and revert to using random seeds.